refactor(shorebird_cli): throw error if incomplete engine config args are provided
